+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - GitLab: [#1386] Move from `@gitbeaker/node` to `@gitbeaker/rest` [@heltoft]
 - GitLab: [#1412] Danger fails to create inline comments on Gitlab [@heltoft]
 - GitLab: [#1405] Can't post multiple inline comments [@heltoft]
+- GitLab: Do not delete system resolved danger inline comments [@heltoft]
 <!-- Your comment above this -->
 
 ## 11.3.1